+++ src/Symfony/Component/Messenger/Tests/DependencyInjection/MessengerBundleExtensionTest.php
@@ -35,12 +35,19 @@ use Symfony\Component\Messenger\Bridge\Beanstalkd\Transport\BeanstalkdTransportF
 use Symfony\Component\Messenger\Bridge\MongoDb\Transport\MongoDbTransportFactory;
 use Symfony\Component\Messenger\Bridge\Redis\Transport\RedisTransportFactory;
 use Symfony\Component\Messenger\DependencyInjection\MessengerPass;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\MessengerBundle;
+use Symfony\Component\Messenger\Middleware\AddIdentityStampsMiddleware;
+use Symfony\Component\Messenger\Middleware\PropagateStampsMiddleware;
+use Symfony\Component\Messenger\Middleware\StackMiddleware;
+use Symfony\Component\Messenger\Stamp\MessageIdStamp;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Transport\Sender\OutboxSender;
 use Symfony\Component\Messenger\Transport\Serialization\ClaimCheckSerializer;
 use Symfony\Component\Messenger\Transport\TransportFactory;
+use Symfony\Component\Uid\Uuid;
+use Symfony\Component\Uid\UuidV7;
 
 class MessengerBundleExtensionTest extends TestCase
 {
@@ -179,6 +186,57 @@ class MessengerBundleExtensionTest extends TestCase
 
         $this->assertNotContains('messenger.middleware.reject_redelivered_message_middleware', $this->getBusMiddlewareIds($container, 'messenger.bus.default'));
         $this->assertContains('messenger.middleware.reject_redelivered_message_middleware', $this->getBusMiddlewareIds($container, 'messenger.bus.commands'));
+    }
+
+    public function testMessengerIdentityStampsAreDisabledByDefault()
+    {
+        $container = $this->createContainerFromFile('messenger', false);
+        $container->compile();
+
+        $this->assertNotContains(['id' => 'add_identity_stamps'], $container->getParameter('messenger.bus.default.middleware'));
+        $this->assertFalse($container->hasDefinition('messenger.middleware.add_identity_stamps'));
+        $this->assertFalse($container->hasDefinition('messenger.message_id_generator'));
+    }
+
+    public function testMessengerIdentityStampsMiddlewareRunsBeforePropagateStamps()
+    {
+        $container = $this->createContainerFromFile('messenger_identity_stamps', false);
+        $container->compile();
+
+        $middleware = array_values($container->getParameter('messenger.bus.default.middleware'));
+        $position = array_search(['id' => 'add_identity_stamps'], $middleware, true);
+
+        $this->assertNotFalse($position, 'The add_identity_stamps middleware is listed.');
+        $this->assertSame(['id' => 'propagate_stamps'], $middleware[$position + 1]);
+        $this->assertSame(AddIdentityStampsMiddleware::class, $container->getDefinition('messenger.middleware.add_identity_stamps')->getClass());
+    }
+
+    public function testMessengerIdentityStampsUseUuidV7WhenTheUidComponentIsInstalled()
+    {
+        if (!class_exists(Uuid::class)) {
+            $this->markTestSkipped('The Uid component is not installed.');
+        }
+
+        $container = $this->createContainerFromFile('messenger_identity_stamps', false);
+        $container->register('foo', \stdClass::class)
+            ->setPublic(true)
+            ->setProperty('middleware', new Reference('messenger.middleware.add_identity_stamps'));
+        $container->compile();
+
+        $envelope = $container->get('foo')->middleware->handle(new Envelope(new \stdClass()), new StackMiddleware());
+
+        $this->assertInstanceOf(UuidV7::class, Uuid::fromString($envelope->last(MessageIdStamp::class)->getId()));
+    }
+
+    public function testMessengerPropagateStampsMiddlewareIsSharedByAllBuses()
+    {
+        $container = $this->createContainerFromFile('messenger_reject_redelivered_messages_disabled_explicit_bus', false);
+        $container->addCompilerPass(new MessengerPass());
+        $container->compile();
+
+        $this->assertSame(PropagateStampsMiddleware::class, $container->getDefinition('messenger.middleware.propagate_stamps')->getClass());
+        $this->assertContains('messenger.middleware.propagate_stamps', $this->getBusMiddlewareIds($container, 'messenger.bus.default'));
+        $this->assertContains('messenger.middleware.propagate_stamps', $this->getBusMiddlewareIds($container, 'messenger.bus.commands'));
     }
 
     public function testMessengerMultipleFailureTransports()
@@ -525,6 +583,7 @@ class MessengerBundleExtensionTest extends TestCase
         $this->assertEquals([
             ['id' => 'add_default_stamps_middleware'],
             ['id' => 'add_bus_name_stamp_middleware', 'arguments' => ['messenger.bus.events']],
+            ['id' => 'propagate_stamps'],
             ['id' => 'reject_redelivered_message_middleware'],
             ['id' => 'dispatch_after_current_bus'],
             ['id' => 'decode_failed_message_middleware'],
@@ -544,6 +603,7 @@ class MessengerBundleExtensionTest extends TestCase
         $this->assertEquals([
             ['id' => 'add_default_stamps_middleware'],
             ['id' => 'add_bus_name_stamp_middleware', 'arguments' => ['messenger.bus.commands']],
+            ['id' => 'propagate_stamps'],
             ['id' => 'reject_redelivered_message_middleware'],
             ['id' => 'dispatch_after_current_bus'],
             ['id' => 'decode_failed_message_middleware'],
@@ -557,6 +617,7 @@ class MessengerBundleExtensionTest extends TestCase
         $this->assertEquals([
             ['id' => 'add_default_stamps_middleware'],
             ['id' => 'add_bus_name_stamp_middleware', 'arguments' => ['messenger.bus.events']],
+            ['id' => 'propagate_stamps'],
             ['id' => 'reject_redelivered_message_middleware'],
             ['id' => 'dispatch_after_current_bus'],
             ['id' => 'decode_failed_message_middleware'],